+++ packages/cli/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the Zowe CLI package will be documented in this file.
 
 ## Recent Changes
 
+-Enhancement: Added new command zowe zos-files download all-members-matching, (zowe files dl amm), to download members matching specified pattern(s). [#2175](https://github.com/zowe/zowe-cli/issues/2175)
+
+## Recent Changes
+
 - Enhancement: Added --wait-for-active and --wait-for-output to download options on zosjobs. [#2328](https://github.com/zowe/zowe-cli/pull/2328)
 
 ## `8.6.0`